refactor: structure of annotation generator
